Conflicts:
	.gitignore
	CMakeLists.txt
	cmake/LocalConfiguration.cmake
	cmake/Version.cmake
	debian/changelog
	library/json/CMakeLists.txt
	library/lfalloc/CMakeLists.txt
	library/lfalloc/helpers.h
	library/lfalloc/lf_allocX64.cpp
	python/Makefile
	python/yt/bindings/CMakeLists.txt
	python/yt/bindings/driver.cpp
	python/yt/bindings/gil.h
	python/yt/bindings/stream.cpp
	python/yt/environment/configs.py
	python/yt/environment/yt_env.py
	python/yt/wrapper/py_wrapper.py
	scripts/dashboard-kant.html
	scripts/mirrors-setup.sh
	tests/integration/tests/test_cypress_commands.py
	tests/integration/tests/test_other.py
	tests/integration/tests/test_table_commands.py
	tests/integration/yt_commands.py
	tests/integration/yt_env_setup.py
	util
	yt/core/CMakeLists.txt
	yt/core/actions/bind.h.pump
	yt/core/actions/callback.h
	yt/core/actions/callback_internal.h
	yt/core/actions/signal.h
	yt/core/bus/tcp_dispatcher.h
	yt/core/compression/codec.cpp
	yt/core/compression/helpers.h
	yt/core/concurrency/action_queue.cpp
	yt/core/concurrency/action_queue_detail.cpp
	yt/core/concurrency/action_queue_detail.h
	yt/core/concurrency/delayed_executor.cpp
	yt/core/concurrency/fiber.cpp
	yt/core/concurrency/parallel_collector-inl.h
	yt/core/concurrency/periodic_executor.cpp
	yt/core/concurrency/public.h
	yt/core/concurrency/thread_affinity.h
	yt/core/formats/helpers.cpp
	yt/core/formats/schemaful_dsv_parser.cpp
	yt/core/formats/schemaful_dsv_parser.h
	yt/core/formats/schemaful_dsv_table.h
	yt/core/formats/schemaful_dsv_writer.cpp
	yt/core/formats/schemaful_dsv_writer.h
	yt/core/formats/schemed_dsv_parser.cpp
	yt/core/formats/schemed_dsv_parser.h
	yt/core/formats/schemed_dsv_table.h
	yt/core/formats/schemed_dsv_writer.h
	yt/core/logging/common.h
	yt/core/logging/log_manager.cpp
	yt/core/logging/writer.cpp
	yt/core/logging/writer.h
	yt/core/misc/CMakeLists.txt
	yt/core/misc/chunked_memory_pool.cpp
	yt/core/misc/chunked_memory_pool.h
	yt/core/misc/error.h
	yt/core/misc/fs.cpp
	yt/core/misc/home.cpp
	yt/core/misc/lazy_holder.h
	yt/core/misc/object_pool-inl.h
	yt/core/misc/object_pool.h
	yt/core/misc/serialize.h
	yt/core/profiling/profiler.cpp
	yt/core/rpc/bus_channel.cpp
	yt/core/rpc/config.h
	yt/core/yson/detail.h
	yt/core/yson/parser.cpp
	yt/core/yson/parser.h
	yt/core/yson/parser_detail.h
	yt/core/yson/varint.cpp
	yt/core/ytree/fluent.h
	yt/core/ytree/node_detail.cpp
	yt/core/ytree/yson_serializable.cpp
	yt/driver/main.cpp
	yt/experiments/CMakeLists.txt
	yt/experiments/benchmark_formats/main.cpp
	yt/experiments/decompress_snapshot/CMakeLists.txt
	yt/experiments/election/CMakeLists.txt
	yt/experiments/fibers/CMakeLists.txt
	yt/experiments/hydra_test/CMakeLists.txt
	yt/experiments/messagebus_client/CMakeLists.txt
	yt/experiments/messagebus_server/CMakeLists.txt
	yt/experiments/messagebus_server/config.h
	yt/experiments/messagebus_server/main.cpp
	yt/nodejs/lib/command.js
	yt/nodejs/src/common.cpp
	yt/nodejs/src/node.cpp
	yt/server/cell_master/meta_state_facade.cpp
	yt/server/cell_master/serialize.cpp
	yt/server/cell_node/bootstrap.cpp
	yt/server/cell_node/bootstrap.h
	yt/server/cell_scheduler/bootstrap.h
	yt/server/cell_scheduler/config.h
	yt/server/chunk_server/chunk_list_proxy.cpp
	yt/server/chunk_server/chunk_owner_node_proxy.cpp
	yt/server/chunk_server/chunk_placement.cpp
	yt/server/chunk_server/chunk_proxy.cpp
	yt/server/chunk_server/helpers-inl.h
	yt/server/data_node/data_node_service.cpp
	yt/server/data_node/job.cpp
	yt/server/data_node/master_connector.cpp
	yt/server/data_node/session_manager.cpp
	yt/server/job_proxy/job_proxy.cpp
	yt/server/job_proxy/partition_sort_job.cpp
	yt/server/job_proxy/user_job.cpp
	yt/server/job_proxy/user_job_io.cpp
	yt/server/main.cpp
	yt/server/misc/config.h
	yt/server/node_tracker_server/node.cpp
	yt/server/orchid/cypress_integration.cpp
	yt/server/scheduler/CMakeLists.txt
	yt/server/scheduler/chunk_info_collector-inl.h
	yt/server/scheduler/chunk_splits_fetcher.cpp
	yt/server/scheduler/config.h
	yt/server/scheduler/fair_share_strategy.cpp
	yt/server/scheduler/helpers.cpp
	yt/server/scheduler/map_controller.cpp
	yt/server/scheduler/master_connector.cpp
	yt/server/scheduler/merge_controller.cpp
	yt/server/scheduler/operation_controller.h
	yt/server/scheduler/operation_controller_detail.cpp
	yt/server/scheduler/operation_controller_detail.h
	yt/server/scheduler/samples_fetcher.cpp
	yt/server/scheduler/scheduler.cpp
	yt/server/scheduler/scheduler_service.cpp
	yt/server/scheduler/snapshot_builder.cpp
	yt/server/scheduler/snapshot_downloader.cpp
	yt/server/scheduler/sort_controller.cpp
	yt/server/table_server/table_node_proxy.cpp
	yt/unittests/fls_ut.cpp
	yt/unittests/heap_ut.cpp
	yt/unittests/json_parser_ut.cpp
	yt/unittests/json_writer_ut.cpp
	yt/unittests/main.cpp
	yt/unittests/schemaful_dsv_parser_ut.cpp
	yt/unittests/schemaful_dsv_writer_ut.cpp
	yt/unittests/yamr_writer_ut.cpp
	yt/unittests/yson_serializable_ut.cpp
	yt/ytlib/CMakeLists.txt
	yt/ytlib/chunk_client/chunk_spec.cpp
	yt/ytlib/chunk_client/config.h
	yt/ytlib/chunk_client/erasure_reader.cpp
	yt/ytlib/chunk_client/erasure_writer.cpp
	yt/ytlib/chunk_client/key.h
	yt/ytlib/chunk_client/multi_chunk_sequential_writer-inl.h
	yt/ytlib/chunk_client/replication_reader.cpp
	yt/ytlib/chunk_client/replication_writer.cpp
	yt/ytlib/chunk_client/sequential_reader.cpp
	yt/ytlib/driver/command.h
	yt/ytlib/driver/config.h
	yt/ytlib/driver/cypress_commands.cpp
	yt/ytlib/driver/driver.cpp
	yt/ytlib/driver/etc_commands.cpp
	yt/ytlib/driver/table_commands.cpp
	yt/ytlib/driver/transaction_commands.cpp
	yt/ytlib/election/config.h
	yt/ytlib/file_client/file_chunk_output.cpp
	yt/ytlib/file_client/file_reader.cpp
	yt/ytlib/file_client/file_writer.cpp
	yt/ytlib/file_client/private.cpp
	yt/ytlib/file_client/private.h
	yt/ytlib/formats/config.h
	yt/ytlib/formats/format.cpp
	yt/ytlib/formats/helpers.h
	yt/ytlib/formats/json_parser.cpp
	yt/ytlib/formats/json_parser.h
	yt/ytlib/formats/json_writer.cpp
	yt/ytlib/formats/schemed_dsv_parser.cpp
	yt/ytlib/formats/schemed_dsv_parser.h
	yt/ytlib/formats/schemed_dsv_table.h
	yt/ytlib/formats/schemed_dsv_writer.h
	yt/ytlib/meta_state/change_log_cache.cpp
	yt/ytlib/meta_state/config.h
	yt/ytlib/meta_state/decorated_meta_state.cpp
	yt/ytlib/meta_state/map.h
	yt/ytlib/meta_state/master_channel.cpp
	yt/ytlib/meta_state/master_discovery.cpp
	yt/ytlib/meta_state/master_discovery.h
	yt/ytlib/meta_state/persistent_state_manager.cpp
	yt/ytlib/meta_state/recovery.cpp
	yt/ytlib/meta_state/snapshot_builder.cpp
	yt/ytlib/node_tracker_client/node.proto
	yt/ytlib/node_tracker_client/node_directory.h
	yt/ytlib/orchid/orchid_service.cpp
	yt/ytlib/orchid/orchid_service.h
	yt/ytlib/orchid/orchid_service_proxy.h
	yt/ytlib/scheduler/config.h
	yt/ytlib/scheduler/public.h
	yt/ytlib/shutdown.cpp
	yt/ytlib/table_client/partition_chunk_writer.cpp
	yt/ytlib/table_client/private.cpp
	yt/ytlib/table_client/private.h
	yt/ytlib/table_client/public.h
	yt/ytlib/table_client/table_chunk_reader.cpp
	yt/ytlib/table_client/table_writer.cpp
	yt/ytlib/table_client/table_writer.h
	yt/ytlib/transaction_client/transaction_manager.cpp

+++ yt/ytlib/table_client/table_chunk_writer.cpp
@@ -182,7 +182,7 @@ void TTableChunkWriter::FinalizeRow(const TRow& row)
         CurrentUncompressedSize += channel->GetDataSize();
     }
 
-    CurrentSize = static_cast<int>(EncodingWriter->GetCompressionRatio() * CurrentUncompressedSize);
+    CurrentSize = static_cast<i64>(EncodingWriter->GetCompressionRatio() * CurrentUncompressedSize);
 
     while (BuffersHeap.front()->GetDataSize() > static_cast<size_t>(Config->BlockSize)) {
         PrepareBlock();
